fix(candle-util): run on Metal — no F64 device tensors, and a backend fallback for the row gather ops
`senna bge --device metal` died twice on the way through a fit. sparsemax
built its 1..k ramp with f64 endpoints, so the ramp landed on the device
as F64 and the cast to the working dtype hit Metal's missing F64→F32
conversion; the two `Tensor::full` fills in the checkpoint-growth path
had the same shape of bug. All three now build in f32.

The custom `gather_rows` / `index_add_rows` ops have CPU and CUDA
kernels only, so on Metal they now fall back to candle's own
`index_select` / `index_add`, which that backend implements. The
row-parallel backward is a CUDA optimisation and nothing on Metal
depends on it.

candle-util 0.8.4 → 0.8.5.

+++ candle-util/src/nn/layers.rs
@@ -226,7 +226,7 @@ pub fn sparsemax(z: &Tensor) -> Result<Tensor> {
     let cumsum = z_sorted.cumsum(dim)?;
 
     // Compute 1 + i * z_sorted[i] for i = 1..k
-    let range = Tensor::arange(1.0, (k + 1) as f64, device)?.to_dtype(dtype)?;
+    let range = Tensor::arange(1f32, (k + 1) as f32, device)?.to_dtype(dtype)?;
     // Broadcast range to match z shape
     let shape: Vec<usize> = (0..z.rank())
         .map(|i| if i == dim { k } else { 1 })
